+++ plugin_ide.ui/src/com/github/rustdt/ide/ui/editor/RustSourceViewerConfiguration.java
@@ -96,11 +96,6 @@ public class RustSourceViewerConfiguration extends AbstractLangSourceViewerConfi
 			protected ILangCompletionProposalComputer createDefaultSymbolsProposalComputer() {
 				return new RustCompletionProposalComputer();
 			}
-			
-			@Override
-			protected ILangCompletionProposalComputer createSnippetsProposalComputer() {
-				return null; // TODO 
-			}
 		};
 	}