+++ plugin_ide.ui/src/melnorme/lang/ide/ui/EditorSettings_Actual.java
@@ -10,11 +10,17 @@
  *******************************************************************************/
 package melnorme.lang.ide.ui;
 
+import static melnorme.utilbox.core.CoreUtil.array;
+import melnorme.lang.ide.core.TextSettings_Actual.LangPartitionTypes;
 import melnorme.lang.ide.ui.editor.LangEditorContextMenuContributor;
 import melnorme.lang.ide.ui.editor.text.EditorPrefConstants_Common;
 
 import org.eclipse.cdt.ui.text.IColorManager;
 import org.eclipse.jface.preference.IPreferenceStore;
+import org.eclipse.jface.text.contentassist.ContentAssistant;
+import org.eclipse.jface.text.contentassist.IContentAssistProcessor;
+import org.eclipse.jface.text.source.ISourceViewer;
+import org.eclipse.jface.text.source.SourceViewerConfiguration;
 import org.eclipse.ui.services.IServiceLocator;
 
 import com.github.rustdt.ide.ui.actions.RustEditorContextMenuContributor;
@@ -45,7 +51,27 @@ public class EditorSettings_Actual {
 	public static final String TEMPLATE_CONTEXT_TYPE_ID = LangUIPlugin.PLUGIN_ID + ".TemplateContextType";
 	
 	public static final String CODE_DEFAULT_COLOR = RustColorPreferences.DEFAULT.key;
-		
+	
+	public static SourceViewerConfiguration createTemplateEditorSourceViewerConfiguration(
+			IPreferenceStore store, final IContentAssistProcessor templateCAP) {
+		IColorManager colorManager = LangUIPlugin.getInstance().getColorManager();
+		return new RustSimpleSourceViewerConfiguration(store, colorManager) {
+			@Override
+			public ContentAssistant getContentAssistant(ISourceViewer sourceViewer) {
+				return setupSimpleContentAssistant(templateCAP, array(
+					LangPartitionTypes.CODE.getId(), 
+					LangPartitionTypes.LINE_COMMENT.getId(), 
+					LangPartitionTypes.BLOCK_COMMENT.getId(), 
+					LangPartitionTypes.DOC_LINE_COMMENT.getId(),
+					LangPartitionTypes.DOC_BLOCK_COMMENT.getId(), 
+					LangPartitionTypes.LIFETIME.getId(),
+					LangPartitionTypes.ATTRIBUTE.getId()
+				));
+			}
+			
+		};
+	}
+	
 	/* ----------------- actions ----------------- */
 	
 	public static interface EditorCommandIds {